Normalize line endings
Apparently all files were commited with LF line endings already, so no other files had to be changed as part of this. But the nice part about being explicit like this is now it'll behave on every git clone rather than depend on the machine settings on each clone to match the author's.